Move precision parsing test to test_types
