+++ test/pytest/test_types.py
@@ -1,3 +1,6 @@
+import pytest
+
+from hls4ml.backends.fpga.fpga_backend import FPGABackend
 from hls4ml.backends.fpga.fpga_types import ACFixedPrecisionDefinition, APFixedPrecisionDefinition
 from hls4ml.model.types import (
     ExponentPrecisionType,
@@ -55,3 +58,29 @@ def test_precision_type_creation():
     fp_type.saturation_bits = 0
     fp_cpp = fp_type.definition_cpp()
     assert fp_cpp == 'ac_fixed<12,6,true>'  # Should not include defaults
+
+
+@pytest.mark.parametrize(
+    'prec_pair',
+    [
+        ('ap_fixed<3, 2>', True),
+        ('ap_ufixed<3, 2>', False),
+        ('ac_fixed<3, 2, true>', True),
+        ('ac_fixed<3, 2, false>', False),
+        ('ac_fixed<3, 2, 1>', True),
+        ('ac_fixed<3, 2, 0>', False),
+        ('ap_int<3, 2>', True),
+        ('ap_uint<3>', False),
+        ('ac_int<3, TRue>', True),
+        ('ac_int<3, FALse>', False),
+        ('ac_int<3, 1>', True),
+        ('ac_int<3, 0>', False),
+    ],
+)
+def test_sign_parsing(prec_pair):
+    '''Test that convert_precisions_string determines the signedness correctly'''
+    strprec = prec_pair[0]
+    signed = prec_pair[1]
+
+    evalprec = FPGABackend.convert_precision_string(strprec)
+    assert evalprec.signed == signed
